@@ -1,5 +1,3 @@
-import { NextPage } from "next";
-import React from "react";
 import AboutMe from "@/components/AboutMe";
 import Career from "@/components/Career";
 import Award from "@/components/Award";
@@ -10,7 +8,7 @@ const classes = {
   homeWrap: "grid grid-cols-1 gap-y-12",
 };
 
-const Home: NextPage = () => {
+export default function Home() {
   return (
     <div className={classes.homeWrap}>
       <AboutMe />
@@ -20,6 +18,4 @@ const Home: NextPage = () => {
       <Others />
     </div>
   );
-};
-
-export default Home;
+}
