@@ -1,15 +1,11 @@
-import React, { ReactNode } from "react";
+import type { ReactNode } from "react";
 
 type Props = {
   children?: ReactNode;
 };
 
-const Layout = ({ children }: Props) => (
+const Layout = ({ children = null }: Props) => (
   <div className="px-10 mt-8 mb-16 max-w-7xl py-8 lg:px-16">{children}</div>
 );
-
-Layout.defaultProps = {
-  children: null,
-};
 
 export default Layout;
